Added write to boot tracks
